Move one global to an extern, since otherwise clang dies (and it was technically an error anyway...)

+++ pyBigWig.c
@@ -153,8 +153,8 @@ static PyObject *pyBwGetStats(pyBigWigFile_t *self, PyObject *args, PyObject *kw
         Py_INCREF(Py_None);
         return Py_None;
     }
-    start = startl;
-    end = endl;
+    start = (uint32_t) startl;
+    end = (uint32_t) endl;
     if(end <= start || end > bw->cl->len[tid] || start >= end) {
         PyErr_SetString(PyExc_RuntimeError, "Invalid interval bounds!");
         return NULL;
@@ -202,8 +202,8 @@ static PyObject *pyBwGetValues(pyBigWigFile_t *self, PyObject *args) {
         Py_INCREF(Py_None);
         return Py_None;
     }
-    start = startl;
-    end = endl;
+    start = (uint32_t) startl;
+    end = (uint32_t) endl;
     if(end <= start || end > bw->cl->len[tid] || start >= end) {
         PyErr_SetString(PyExc_RuntimeError, "Invalid interval bounds!");
         return NULL;
@@ -244,8 +244,8 @@ static PyObject *pyBwGetIntervals(pyBigWigFile_t *self, PyObject *args, PyObject
         Py_INCREF(Py_None);
         return Py_None;
     }
-    start = startl;
-    end = endl;
+    start = (uint32_t) startl;
+    end = (uint32_t) endl;
     if(end <= start || end > bw->cl->len[tid] || start >= end) {
         PyErr_SetString(PyExc_RuntimeError, "Invalid interval bounds!");
         return NULL;
